Add lucene regex rewriting
The lucene regex syntax is lacking some basic features that users
expect from regex engines.  Some of these features can be added
by rewriting the regex, along with transforming the input value.

Supported features:
* Minimal set of shorthand character classes: \d, \s, and \w
* Start/end anchors: ^ and $

The transformation code is being put in the utils module so it
can be reused in both the opensearch-extra and the
opensearch-cirrus-highlighter plugins.

Bug: T317599
Change-Id: Ie5e3ad991235fdd0b6efffc9c747b406e52e8e00