regex: Support quoted literals
Lucene regexp's include a quoted literal syntax, but the regex
rewriter was not taking these into account. Add test cases
asserting appropriate expansions, and implement the necessary
changes to pass them.

This only tests the expansions directly, and not the regex
equivalence, because the java Pattern we use to assert
equivalence does not support the same quoted literals.

Bug: T399162
Change-Id: Ibce4461cfa489b1718bf9e7914f925ee29b204e1